+++ apps/ui-dev/pages/Sample.jsx
@@ -1,4 +1,4 @@
-import { useContext, useEffect, useState } from "@elf-framework/sapa";
+import { useContext, useState } from "@elf-framework/sapa";
 
 import { MyContext } from "./context";
 
@@ -9,21 +9,11 @@ export { B };
 
 function useMyState(initialValue) {
   const [value, setValue] = useState(initialValue);
-
-  useEffect(() => {
-    console.log("use effected value", value);
-
-    return () => {
-      console.log("unmount");
-    };
-  }, []);
-
   return [value, setValue];
 }
 
 export default function Sample({ onClick }) {
   const myContextValue = useContext(MyContext);
-  // console.log({ myContextValue });
   const [value, setValue] = useMyState(1);
   const [value2, setValue2] = useState(10);
 
@@ -40,7 +30,7 @@ export default function Sample({ onClick }) {
             this.runCallback(onClick);
           }}
         >
-          fdsafdsafds af {value + 10 ** 2} yello {myContextValue.value}
+          fdsafdsafds af {value + 10 ** 2} blue 20 {myContextValue.value}
         </button>
       )}
       <button