+++ packages/base-editor/src/index.js
@@ -1,3 +1,4 @@
-import "@elf-framework/ui/style.css?raw";
+import "@elf-framework/ui/style.css";
 
+export * from "./components/editor/Editor";
 export * from "./components/editor/BaseEditor";