+++ packages/base-editor/src/managers/UIManager.js
@@ -1,0 +1,14 @@
+export class UIManager {
+  constructor(editorContext) {
+    this.editorContext = editorContext;
+    this.uis = {};
+  }
+
+  registerUI(obj = {}) {
+    Object.assign(this.uis, obj);
+  }
+
+  getUI(key) {
+    return this.uis[key];
+  }
+}